requirements.txt
aiohttp==1.0.5
alembic==0.9.6
async-timeout==2.0.0
banal==0.3.3
certifi==2017.11.5
chardet==3.0.4
dataset==1.0.5
discord.py==0.16.12
idna==2.6
Mako==1.0.7
MarkupSafe==1.0
multidict==4.0.0
normality==0.5.7
PyMySQL==0.8.0
python-dateutil==2.6.1
python-editor==1.0.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.0
urllib3==1.22
utopian-api==0.0.1
websockets==3.4